// src/app/interface/register-dto.ts
export interface RegisterDTO {
  name?: string;
  email?: string;
  provider?: string;
  password?: string;
  idProvider?: string;
  photoUrl?: string;
  firstName?: string;
  lastName?: string;
}


// provider?: string;
//     id?: string;
//     email?: string;
//     name?: string;
//     photoUrl?: string;
//     firstName?: string;
//     lastName?: string;
//     authToken?: string;